In case additional condition information
shall be written to the output file of the dbc monitor,
a yaml file needs to be written.
This yaml output is added as an alternative.
A test is added to test the new output functionality
The file comparison python script can now compare both csv files
and yaml files based on code from FourCIPP